Create 0x3e98ebdd41b52e62bd984e070cfce2031114eaf9
NGL is a model of the integration of Web3 and RWA technologies, and is the basic token of the original ecology of the global farming chain; it is mainly responsible for promoting the flow and exchange of value in the entire farming chain ecosystem. It can be used to purchase agricultural services, pledge to obtain NGFS, obtain data and resources, participate in ecological governance, etc. The value stability and liquidity of NGL have become an ideal choice for transactions between agricultural producers, suppliers, consumers and other participants; it assumes an important system role in the ecological value delivery and governance of the global farming chain.